Refactor BestillingMalService and update related classes
A significant refactor on BestillingMalService has been conducted by moving the logic within MalBestillingService to BestillingMalService. As a result, MalBestillingService is not needed anymore and it was removed. Alongside, the related classes and tests were updated according to the new structure.
