+++ apps/dolly-backend/src/main/java/no/nav/dolly/provider/api/MalBestillingController.java
@@ -1,25 +1,74 @@
 package no.nav.dolly.provider.api;
 
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import no.nav.dolly.domain.resultset.entity.bestilling.RsBestillingMal;
-import no.nav.dolly.service.MalBestillingService;
+import no.nav.dolly.domain.resultset.entity.bestilling.RsMalBestillingWrapper;
+import no.nav.dolly.service.BestillingMalService;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import static no.nav.dolly.config.CachingConfig.CACHE_BESTILLING_MAL;
+import static org.apache.commons.lang3.StringUtils.isBlank;
 
 @RequestMapping(value = "/api/v1/malbestilling")
 @RestController
 @RequiredArgsConstructor
 public class MalBestillingController {
 
-    private final MalBestillingService malBestillingService;
+    private final BestillingMalService bestillingMalService;
 
+    @CacheEvict(value = { CACHE_BESTILLING_MAL }, allEntries = true)
     @PostMapping(value = "/ident/{ident}")
+    @Operation(description = "Opprett ny mal-bestilling fra ident")
     public RsBestillingMal createTemplateFromIdent(@PathVariable String ident,
-                                                   @RequestParam String name) {
+                                                   @RequestParam String malNavn) {
 
-        return malBestillingService.createFromIdent(ident, name);
+        return bestillingMalService.createFromIdent(ident, malNavn);
+    }
+
+    @Cacheable(value = CACHE_BESTILLING_MAL)
+    @GetMapping
+    @Operation(description = "Hent mal-bestilling, kan filtreres på en bruker")
+    public RsMalBestillingWrapper getMalBestillinger(@RequestParam(required = false) String brukerId) {
+
+        return isBlank(brukerId) ?
+                bestillingMalService.getMalBestillinger() : bestillingMalService.getMalbestillingByUser(brukerId);
+    }
+
+    @CacheEvict(value = { CACHE_BESTILLING_MAL }, allEntries = true)
+    @PostMapping
+    @Operation(description = "Opprett ny mal-bestilling fra bestillingId")
+    @Transactional
+    public void opprettMalbestilling(@RequestParam Long bestillingId, @RequestParam String malNavn) {
+
+        bestillingMalService.saveBestillingMalFromBestillingId(bestillingId, malNavn);
+    }
+
+    @CacheEvict(value = { CACHE_BESTILLING_MAL }, allEntries = true)
+    @DeleteMapping("/id/{id}")
+    @Operation(description = "Slett mal-bestilling")
+    @Transactional
+    public void deleteMalBestilling(@PathVariable Long id) {
+
+        bestillingMalService.deleteMalBestillingByID(id);
+    }
+
+    @CacheEvict(value = { CACHE_BESTILLING_MAL }, allEntries = true)
+    @PutMapping("/id/{id}")
+    @Operation(description = "Rediger mal-bestilling")
+    @Transactional
+    public void redigerMalBestilling(@PathVariable Long id, @RequestParam String malNavn) {
+
+        bestillingMalService.updateMalNavnById(id, malNavn);
     }
 }
